Moved html files to views directory
