+++ src/constants.json
@@ -1,1 +1,16 @@
-{"dataSendRate":30,"renderInterval":250,"serverAddr":{"ip":"0.0.0.0","port":33333},"hvBone":{"ip":"192.168.0.7","port":9094},"lvBone":{"ip":"192.168.0.6","port":9091}}
+{
+    "dataSendRate": 30,
+    "renderInterval": 250,
+    "serverAddr": {
+        "ip": "0.0.0.0",
+        "port": 33333
+    },
+    "hvBone": {
+        "ip": "192.168.0.7",
+        "port": 9094
+    },
+    "lvBone": {
+        "ip": "192.168.0.6",
+        "port": 9091
+    }
+}
