+++ src/views/testing.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <link rel="stylesheet" type="text/css" href="public/stylesheets/main.css">
-  <link rel="stylesheet" type="text/css" href="public/stylesheets/testingtool.grid.css">
-  <link rel="stylesheet" type="text/css" href="public/stylesheets/fonts.css">
-  <script src="public/javascripts/chartHandling.js"></script>
+  <link rel="stylesheet" type="text/css" href="../public/stylesheets/main.css">
+  <link rel="stylesheet" type="text/css" href="../public/stylesheets/testingtool.grid.css">
+  <link rel="stylesheet" type="text/css" href="../public/stylesheets/fonts.css">
+  <script src="../public/javascripts/chartHandling.js"></script>
   <script src="../node_modules/plotly.js-dist/plotly.js"></script>
   <title>Badgerloop PRAT</title>
 </head>
@@ -12,7 +12,7 @@
 <body>
   <!-- Top Bar -->
   <div class="banner">
-    <a class="left"><img src="public/images/logo.png" height="30px"></a>
+    <a class="left"><img src="../public/images/logo.png" height="30px"></a>
     <a id="title" class="left">CONTROL DASHBOARD</a>
     <a id="close-window" class="right close-button">&#10005;</a>
     <a id="max-window" class="right">&#128470; </a>
@@ -22,7 +22,7 @@
         <button class="dropbtn">Change View
           <i class="fa fa-caret-down"></i>
         </button>
-        <img src="public/images/dropdown.png" height="25px">
+        <img src="../public/images/dropdown.png" height="25px">
         <div class="banner_dropdown-content">
           <a href="index.html">Main</a>
           <a href="testing.html">Analysis</a>
@@ -31,7 +31,7 @@
         </div>
       </div>
     </a>
-    <a class="settingsTrigger" id="settingsTrigger"><img src="public/images/settings.png" height="35px"></a>
+    <a class="settingsTrigger" id="settingsTrigger"><img src="../public/images/settings.png" height="35px"></a>
   </div>
 
   <!-- main grid layout-->
@@ -249,8 +249,8 @@
   </div>
   </div>
   </div>
-  <script src="public/javascripts/main.js"></script>
-  <script src="public/javascripts/testingtool.js"></script>
+  <script src="../public/javascripts/main.js"></script>
+  <script src="../public/javascripts/testingtool.js"></script>
 </body>
 
 </html>
